refactor: :truck: move atom/colors -> src/common/constants/colors
