+++ pages/login.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import {Typography} from 'antd';
 
 import LoginSection from '@src/components/common/organisms/sections/login';
-import {GREY, WHITE} from '@src/components/common/atoms/colors';
+import {GREY, WHITE} from '@src/common/constants/colors';
 
 import media from '@src/common/styles/media';
