+++ src/components/common/molecules/button/TableReload.tsx
@@ -1,6 +1,6 @@
 import {Button} from 'antd';
 
-import {INDIGO_BLUE} from '@src/components/common/atoms/colors';
+import {INDIGO_BLUE} from '@src/common/constants/colors';
 import {useBoardContext} from '@src/common/contexts/Board';
 import {ReloadOutlined} from '@ant-design/icons';
